Changed the way titles are done and removed recipe
Used yaml in the head.html to add the Midwest C. etc. to the  page
title instead of coding it into each page title as the previous change
did.

Changed all references to “recipe” to “MCeM2018” to remove the copied
name that means nothing in this context and to setup a name that will
make better sense if it becomes extended.